+++ src/main/java/com/chatty/practice/suhyeon/hello/Recv.java
@@ -1,14 +1,12 @@
-package com.chatty.practice.basic.tutorial.pubandsub;
+package com.chatty.practice.suhyeon.hello;
 
 import com.rabbitmq.client.Channel;
 import com.rabbitmq.client.Connection;
 import com.rabbitmq.client.ConnectionFactory;
 import com.rabbitmq.client.DeliverCallback;
 
-// pub&sub -> queue binding
-public class Subscriber {
-
-    private static final String EXCHANGE_NAME = "logs";
+public class Recv {
+    public  static String QUEUE_NAME = "hello";
 
     public static void main(String[] args) throws Exception {
         ConnectionFactory factory = new ConnectionFactory();
@@ -16,20 +14,13 @@ public class Subscriber {
         Connection connection = factory.newConnection();
         Channel channel = connection.createChannel();
 
-
-        String randomQueue = channel.queueDeclare().getQueue();
-        // last param is binding key. if exchange's type is a fanout, it's ignored
-        channel.queueBind(randomQueue,EXCHANGE_NAME,"");
+        channel.queueDeclare(QUEUE_NAME, false, false, false, null);
+        System.out.println(" [*] Waiting for messages. To exit press CTRL+C");
 
         DeliverCallback deliverCallback = (consumerTag, delivery) -> {
             String message = new String(delivery.getBody(), "UTF-8");
             System.out.println(" [x] Received '" + message + "'");
         };
-
-        boolean autoAck = true;
-        channel.basicConsume(randomQueue, autoAck, deliverCallback, consumerTag -> { });
-        System.out.println(" [*] Waiting for messages. To exit press CTRL+C");
+        channel.basicConsume(QUEUE_NAME, true, deliverCallback, consumerTag -> { });
     }
-
-
 }
