+++ src/main/java/com/chatty/practice/suhyeon/pubandsub/Publisher.java
@@ -1,12 +1,12 @@
-package com.chatty.practice.basic.tutorial.routing;
+package com.chatty.practice.suhyeon.pubandsub;
 
 import com.rabbitmq.client.Channel;
 import com.rabbitmq.client.Connection;
 import com.rabbitmq.client.ConnectionFactory;
 
-public class EmitLogDirect {
+public class Publisher {
 
-    private static final String EXCHANGE_NAME = "direct_logs";
+    private static final String EXCHANGE_NAME = "logs";
 
     public static void main(String[] args) throws Exception {
         ConnectionFactory factory = new ConnectionFactory();
@@ -14,17 +14,18 @@ public class EmitLogDirect {
         try (Connection connection = factory.newConnection();
              Channel channel = connection.createChannel()) {
 
-
-            channel.exchangeDeclare(EXCHANGE_NAME, "direct");
-
-
-            String logMessage = "warning!!";
-            String severity = "warn";
+            channel.exchangeDeclare(EXCHANGE_NAME, "fanout");
 
 
-            channel.basicPublish(EXCHANGE_NAME, severity, null, logMessage.getBytes());
+            String logMessage = "log messages";
+
+            // first param "logs" is a default exchange
+            // don't need a persistence
+            // routeKey "" will be ignored
+            channel.basicPublish(EXCHANGE_NAME, "", null, logMessage.getBytes());
 
             System.out.println(" [x] Sent '" + logMessage + "'");
+
 
 
         }
